Split crates
