+++ Cargo.toml
@@ -1,11 +1,8 @@
-[package]
-name = "ligero"
-version = "0.1.0"
-edition = "2021"
-categories = ["cryptography"]
-license = "MIT OR Apache-2.0"
+[workspace]
+members = ["arithmetic-circuits", "ligero"]
+resolver = "2"
 
-[dependencies]
+[workspace.dependencies]
 ark-std = { version = "0.5.0-alpha", default-features = false }
 ark-ff = { version = "0.5.0-alpha", default-features = false }
 ark-circom = { git = "https://github.com/HungryCatsStudio/circom-compat", branch = "update-0.5" }
@@ -23,17 +20,10 @@ tiny-keccak = { version = "2.0", features = ["keccak"] }
 rand_chacha = { version = "0.3.0", default-features = false }
 blake2 = { version = "0.10", default-features = false }
 
-[dev-dependencies]
-ark-bn254 = { version = "0.5.0-alpha", default-features = false, features = [ "curve" ] }
-ark-bls12-381 = { version = "0.5.0-alpha", default-features = false, features = [ "curve" ] }
-ark-bls12-377 = { version = "0.5.0-alpha", default-features = false, features = [ "curve" ] }
-ark-poly-commit = { git = "https://github.com/HungryCatsStudio/poly-commit", branch = "release-0.5", default-features = false, features = [ "test-types" ] }
-serde = { version = "1.0.104", features = ["derive"] }
-serde_json = "1.0.48"
-criterion = "0.3"
-
-[features]
-default = [ "std" ]
-std = [ "ark-ff/std", "ark-ec/std", "ark-poly/std", "ark-std/std", "ark-relations/std", "ark-serialize/std", "ark-crypto-primitives/std", "ark-poly-commit/std" ]
-test-types = [ "ark-poly-commit/test-types" ]
-bench = [ "std", "test-types" ]
+[workspace.package]
+edition = "2021"
+license = "MIT OR Apache-2.0"
+homepage = "https://github.com/HungryCatsStudio/ligero"
+repository = "https://github.com/HungryCatsStudio/ligero"
+keywords = ["cryptography", "SNARK", "circuits", "ligero"]
+categories = ["cryptography"]
