+++ benches/bench.rs
@@ -1,0 +1,2 @@
+use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use ligero::ligero::LigeroCircuit;
