+++ ligero/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "ligero"
+version = "0.1.0"
+readme = "README.md"
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+ark-std = { workspace = true }
+ark-ff = { workspace = true }
+ark-circom = { workspace = true }
+ark-crypto-primitives = { workspace = true }
+ark-ec = { workspace = true }
+ark-serialize = { workspace = true }
+ark-poly-commit = { workspace = true }
+ark-relations = { workspace = true }
+ark-poly = { workspace = true }
+blake2 = { workspace = true }
+rand_chacha = { workspace = true }
+itertools = { workspace = true }
+
+# Local dependencies
+arithmetic-circuits = { path = "../arithmetic-circuits" , features = [ "examples" ] }
+
+[dev-dependencies]
+ark-bn254 = { version = "0.5.0-alpha", default-features = false, features = [ "curve" ] }
+ark-bls12-381 = { version = "0.5.0-alpha", default-features = false, features = [ "curve" ] }
+ark-bls12-377 = { version = "0.5.0-alpha", default-features = false, features = [ "curve" ] }
+ark-poly-commit = { git = "https://github.com/HungryCatsStudio/poly-commit", branch = "release-0.5", default-features = false, features = [ "test-types" ] }
+serde = { version = "1.0.104", features = ["derive"] }
+serde_json = "1.0.48"
+criterion = "0.3"
+
+[features]
+default = [ "std" ]
+std = [ "ark-ff/std", "ark-ec/std", "ark-poly/std", "ark-std/std", "ark-relations/std", "ark-serialize/std", "ark-crypto-primitives/std", "ark-poly-commit/std" ]
+test-types = [ "ark-poly-commit/test-types" ]
+bench = [ "std", "test-types" ]
